+++ all-is-cubes/src/mesh/block_vertex.rs
@@ -5,28 +5,33 @@ use std::fmt;
 use cgmath::{EuclideanSpace as _, Point3, Vector3};
 
 use crate::math::{Face6, FreeCoordinate, GridPoint, Rgba};
-use crate::util::{ConciseDebug, CustomFormat as _};
+use crate::util::{ConciseDebug, CustomFormat};
 
 /// Numeric type used to store texture coordinates in vertices.
+///
+/// TODO: Delete this type alias now that we're generic over texture coordinates.
 pub type TextureCoordinate = f32;
 
 /// Basic vertex data type from [`triangulate_block`].
 /// Implement <code>[`From`]&lt;[`BlockVertex`]&gt;</code> (and usually [`GfxVertex`])
 /// to provide a specialized version fit for the target graphics API.
 ///
+/// `T` is the type of texture-coordinate points being used. That is, one `T` value
+/// should identify one point in the block's 3D texture, such as `T = Point3<f32>`).
+///
 /// [`triangulate_block`]: super::triangulate_block
 #[allow(clippy::exhaustive_structs)]
 #[derive(Clone, Copy, PartialEq)]
-pub struct BlockVertex {
+pub struct BlockVertex<T> {
     /// Vertex position.
     pub position: Point3<FreeCoordinate>,
     /// Vertex normal, always axis-aligned.
     pub face: Face6,
     /// Surface color or texture coordinate.
-    pub coloring: Coloring,
+    pub coloring: Coloring<T>,
 }
 
-impl BlockVertex {
+impl<T: Clone> BlockVertex<T> {
     /// Remove the clamp information for the sake of tidier tests of one thing at a time.
     #[cfg(test)]
     pub(crate) fn remove_clamps(mut self) -> Self {
@@ -36,9 +41,9 @@ impl BlockVertex {
                 clamp_min: _,
                 clamp_max: _,
             } => Coloring::Texture {
+                clamp_min: pos.clone(),
+                clamp_max: pos.clone(),
                 pos,
-                clamp_min: pos,
-                clamp_max: pos,
             },
             other => other,
         };
@@ -47,26 +52,32 @@ impl BlockVertex {
 }
 
 /// Describes the two ways a [`BlockVertex`] may be colored; by a solid color or by a texture.
+///
+/// `T` is the type of texture-coordinate points being used. That is, one `T` value
+/// should identify one point in the block's 3D texture, such as `T = Point3<f32>`).
 #[allow(clippy::exhaustive_enums)]
-#[derive(Clone, Copy, PartialEq)]
-pub enum Coloring {
+#[derive(Clone, Copy, Eq, PartialEq)]
+pub enum Coloring<T> {
     /// Solid color.
     Solid(Rgba),
     /// Texture coordinates provided by the [`TextureAllocator`](super::TextureAllocator)
     /// for this vertex.
     Texture {
         /// Texture coordinates for this vertex.
-        pos: Vector3<TextureCoordinate>,
+        pos: T,
         /// Lower bounds for clamping the entire surface's texture coordinates.
         /// Used to avoid texture bleed.
-        clamp_min: Vector3<TextureCoordinate>,
+        clamp_min: T,
         /// Upper bounds for clamping the entire surface's texture coordinates.
         /// Used to avoid texture bleed.
-        clamp_max: Vector3<TextureCoordinate>,
+        clamp_max: T,
     },
 }
 
-impl fmt::Debug for BlockVertex {
+impl<T> fmt::Debug for BlockVertex<T>
+where
+    Coloring<T>: fmt::Debug,
+{
     fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
         // Print compactly on single line even if the formatter is in prettyprint mode.
         write!(
@@ -78,7 +89,10 @@ impl fmt::Debug for BlockVertex {
         )
     }
 }
-impl fmt::Debug for Coloring {
+impl<T> fmt::Debug for Coloring<T>
+where
+    T: CustomFormat<ConciseDebug>, // TODO: inelegant
+{
     // TODO: test formatting of this
     fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
@@ -104,7 +118,7 @@ impl fmt::Debug for Coloring {
 /// [`BlockMesh`]: crate::mesh::BlockMesh
 /// [`Space`]: crate::space::Space
 /// [`SpaceMesh`]: crate::mesh::SpaceMesh
-pub trait GfxVertex: From<BlockVertex> + Copy + Sized {
+pub trait GfxVertex: From<BlockVertex<Self::TexPoint>> + Copy + Sized {
     /// Whether [`SpaceMesh`]es should provide pre-sorted vertex index slices to allow
     /// back-to-front drawing order based on viewing ranges.
     ///
@@ -118,6 +132,9 @@ pub trait GfxVertex: From<BlockVertex> + Copy + Sized {
 
     /// Number type for the vertex position coordinates.
     type Coordinate: cgmath::BaseFloat;
+
+    /// Point type identifying a point in the block's texture.
+    type TexPoint: Copy;
 
     /// Type of the data carried from [`Self::instantiate_block`] to
     /// [`Self::instantiate_vertex`].
@@ -139,9 +156,10 @@ pub trait GfxVertex: From<BlockVertex> + Copy + Sized {
 }
 
 /// Trivial implementation of [`GfxVertex`] for testing purposes. Discards lighting.
-impl GfxVertex for BlockVertex {
+impl<T: Copy> GfxVertex for BlockVertex<T> {
     const WANTS_DEPTH_SORTING: bool = true;
     type Coordinate = FreeCoordinate;
+    type TexPoint = T;
     type BlockInst = Vector3<FreeCoordinate>;
 
     fn position(&self) -> Point3<FreeCoordinate> {